Initial Commit, new migrations

+++ db/migrate/20170125020159_create_schools.rb
@@ -3,7 +3,7 @@ class CreateSchools < ActiveRecord::Migration
     create_table :schools do |t|
       t.string :name
       t.string :schid
-      t.float :latitude
+      t.decimal :latitude
       t.decimal :longitude
       t.references :district, index: true, foreign_key: true
